requirements.test.txt
# Strictly for tests
coverage
pytest
pytest-asyncio
pytest-cov
pytest-homeassistant-custom-component